Refactor frequency type from unsigned short to unsigned

+++ src/file.c
@@ -134,7 +134,7 @@ void write_list_to_header(list_node *aux, FILE *file){
 	if(aux != NULL){
 		write_list_to_header(aux->next, file);
 		fputc(aux->tree->chars[0], file);
-		fwrite(&aux->tree->frequency, sizeof(unsigned short), 1, file);
+		fwrite(&aux->tree->frequency, sizeof(unsigned), 1, file);
 	}
 }
 
@@ -143,9 +143,9 @@ list_node *read_list_from_header(FILE *file){
 	char ch;
 
 	while((ch = fgetc(file)) != '\0'){
-		unsigned short freq;
+		unsigned freq;
 
-		fread(&freq, sizeof(unsigned short), 1, file);
+		fread(&freq, sizeof(unsigned), 1, file);
 		add_new_node_to_list(&list, create_leaf(ch, freq));
 	}
