+++ src/tree.c
@@ -6,7 +6,7 @@
 #include "list.h"
 #include "tree.h"
 
-tree_node *create_leaf(char character, unsigned short frequency){
+tree_node *create_leaf(char character, unsigned frequency){
 	tree_node *leaf = malloc(sizeof(tree_node));
 
 	leaf->chars = malloc(LEAF_CHARS_LENGTH);
